+++ examples/getting_started.py
@@ -2,19 +2,14 @@
 
 import ultimate_notion as uno
 
-TOKEN = "secret_YOUR_TOKEN_HERE"
-# Change PAGE_TITLE to the title of your page
 PAGE_TITLE = "Getting Started"
 
-with uno.Session(auth=TOKEN) as notion:
+with uno.Session() as notion:
     page = notion.search_page(PAGE_TITLE).item()
     page.show()
 
 # Alternatively, without a context manager:
-notion = uno.Session.get_or_create(auth=TOKEN)
-# `auth` can be omitted if `NOTION_TOKEN` is set in the environment, e.g.
-# notion = Session.get_or_create()
-# which also works for the context manager
+notion = uno.Session()
 page = notion.search_page(PAGE_TITLE).item()
 page.show()
 notion.close()