+++ tests/adapters/google/test_tasks.py
@@ -1,132 +1,86 @@
-from datetime import datetime, timezone
-from unittest.mock import Mock
+from __future__ import annotations
 
 import pytest
-from pydantic import HttpUrl
 
-from ultimate_notion.adapters.google.tasks import GTask, GTaskList, GTasksClient
-
-
-@pytest.fixture
-def mock_resource():
-    resource = Mock()  # spec=Resource makes no sense since it's a dynamically generated class
-    return resource
+from ultimate_notion.adapters.google import GTasksClient
 
 
 @pytest.fixture
-def mock_gtasklist(mock_resource: Mock):
-    return GTaskList(
-        mock_resource,
-        id='tasklist_id',
-        title='Task List',
-        etag='etag_value',
-        kind='tasks#taskList',
-        updated='2023-12-31T23:24:45.139Z',
-        selfLink='https://example.com/tasklist',
-    )
+def client():
+    return GTasksClient(read_only=False)
 
 
-def test_gtask_init(mock_resource: Mock):
-    data = {
-        'id': 'task_id',
-        'title': 'Task Title',
-        'etag': 'etag_value',
-        'kind': 'tasks#task',
-        'updated': '2023-12-31T23:24:45.139Z',
-        'selfLink': 'https://example.com/task',
-    }
-    task = GTask(mock_resource, **data)
-    assert task.id == data['id']
-    assert task.title == data['title']
-    assert task.etag == data['etag']
-    assert task.kind == data['kind']
-    assert task.updated == datetime.strptime(data['updated'], '%Y-%m-%dT%H:%M:%S.%fZ').replace(tzinfo=timezone.utc)
-    assert task.self_link == HttpUrl(data['selfLink'])
+@pytest.mark.vcr()
+def test_gtask_client(client: GTasksClient):
+    new_list = client.create_tasklist('My new tasklist')
+    assert new_list.title == 'My new tasklist'
+
+    same_list = client.get_tasklist(new_list.id)
+    assert same_list == new_list
+
+    all_lists = client.all_tasklists()
+    assert new_list in all_lists
+
+    new_list.delete()
 
 
-def test_gtasklist_init(mock_resource: Mock):
-    data = {
-        'id': 'tasklist_id',
-        'title': 'Task List Title',
-        'etag': 'etag_value',
-        'kind': 'tasks#taskList',
-        'updated': '2023-12-31T23:24:45.139Z',
-        'selfLink': 'https://example.com/tasklist',
-    }
-    tasklist = GTaskList(mock_resource, **data)
-    assert tasklist.id == data['id']
-    assert tasklist.title == data['title']
-    assert tasklist.etag == data['etag']
-    assert tasklist.kind == data['kind']
-    assert tasklist.updated == datetime.strptime(data['updated'], '%Y-%m-%dT%H:%M:%S.%fZ').replace(tzinfo=timezone.utc)
-    assert tasklist.self_link == HttpUrl(data['selfLink'])
+@pytest.mark.vcr()
+def test_gtask_tasklist(client: GTasksClient):
+    new_list = client.create_tasklist('My new tasklist')
+    new_task = new_list.create_task('My new task')
+    assert new_task.title == 'My new task'
+
+    assert new_task in new_list.all_tasks()
+    new_task.completed = True
+    new_list.clear()
+    assert new_task not in new_list.all_tasks()
+
+    same_list = client.get_tasklist(new_list.id)
+    same_list.title = 'My renamed tasklist'
+    assert new_list == same_list
+    assert new_list.title != same_list.title
+    new_list.reload()
+    assert new_list.title == same_list.title
+
+    new_list.delete()
 
 
-def test_gtasklist_delete(mock_resource: Mock, mock_gtasklist: GTaskList):
-    mock_gtasklist.delete()
-    mock_resource.tasklists.return_value.delete.assert_called_once_with(tasklist='tasklist_id')
+@pytest.mark.vcr()
+def test_gtask_task(client: GTasksClient):
+    new_list = client.create_tasklist('My new tasklist')
+    new_task = new_list.create_task('My new task')
+    assert new_task.title == 'My new task'
+    assert new_task.tasklist == new_list
 
+    same_task = new_list.get_task(new_task.id)
+    same_task.title = 'My renamed task'
+    assert new_task == same_task
+    assert same_task.title != new_task.title
+    new_task.reload()
+    assert new_task.title == same_task.title
 
-def test_gtasklist_get_tasks(mock_resource: Mock):
-    mock_resource.tasks.return_value.list.return_value.execute.return_value = {
-        'items': [
-            {
-                'id': 'task_id',
-                'title': 'Task 1',
-                'etag': 'etag_value',
-                'kind': 'tasks#task',
-                'updated': '2023-12-31T23:24:45.139Z',
-                'selfLink': 'https://example.com/task1',
-            },
-            {
-                'id': 'task_id_2',
-                'title': 'Task 2',
-                'etag': 'etag_value_2',
-                'kind': 'tasks#task',
-                'updated': '2023-12-31T23:24:45.139Z',
-                'selfLink': 'https://example.com/task2',
-            },
-        ]
-    }
-    tasklist = GTaskList(
-        mock_resource,
-        id='tasklist_id',
-        title='Task List',
-        etag='etag_value',
-        kind='tasks#taskList',
-        updated='2023-12-31T23:24:45.139Z',
-        selfLink='https://example.com/tasklist',
-    )
-    tasks = tasklist.all_tasks()
-    assert len(tasks) == 2
-    assert isinstance(tasks[0], GTask)
-    assert isinstance(tasks[1], GTask)
+    new_task.notes = 'My notes'
+    assert new_task.notes == 'My notes'
 
+    new_task.completed = True
+    assert new_task.completed
 
-def test_gtasksclient_all_tasklists(mock_resource: Mock):
-    mock_resource.tasklists.return_value.list.return_value.execute.return_value = {
-        'items': [
-            {
-                'id': 'tasklist_id',
-                'title': 'Task List 1',
-                'etag': 'etag_value',
-                'kind': 'tasks#taskList',
-                'updated': '2023-12-31T23:24:45.139Z',
-                'selfLink': 'https://example.com/tasklist1',
-            },
-            {
-                'id': 'tasklist_id_2',
-                'title': 'Task List 2',
-                'etag': 'etag_value_2',
-                'kind': 'tasks#taskList',
-                'updated': '2023-12-31T23:24:45.139Z',
-                'selfLink': 'https://example.com/tasklist2',
-            },
-        ]
-    }
-    client = GTasksClient()
-    client.resource = mock_resource
-    tasklists = client.all_tasklists()
-    assert len(tasklists) == 2
-    assert isinstance(tasklists[0], GTaskList)
-    assert isinstance(tasklists[1], GTaskList)
+    new_task.position_after(None)
+    assert new_task.position == 0
+    assert new_task.parent is None
+    second_task = new_list.create_task('Second task')
+    second_task.position_after(new_task)
+    assert second_task.position == 1
+
+    parent_task = new_list.create_task('Parent task')
+    assert parent_task.children == []
+    new_task.parent = parent_task
+    assert new_task.parent == parent_task
+    assert parent_task.children == [new_task]
+    new_task.parent = None
+    assert new_task.parent is None
+    assert parent_task.children == []
+
+    new_task.delete()
+    assert new_task.is_deleted
+    new_list.delete()